fix: réfactorisation composants modification plugin commentaire
