+++ admin/src/components/DiscussionThreadItem/index.tsx
@@ -1,28 +1,10 @@
-import { Box, Flex, Typography } from '@strapi/design-system';
+import { Box, Flex } from '@strapi/design-system';
 import { FC, PropsWithChildren } from 'react';
 import styled from 'styled-components';
-import { DiscussionThreadItemActions } from './DiscussionThreadItemActions';
-import { DiscussionThreadItemFooter } from './DiscussionThreadItemFooter';
 import { DiscussionThreadItemProps } from './props';
-
-
-export const DiscussionThreadItemContentTypographyRenderer = styled('div')(() => {
-  return {
-    strong: {
-      fontWeight: 'bold',
-    },
-    i: {
-      fontStyle: 'italic',
-    },
-    em: {
-      fontStyle: 'italic',
-    },
-    u: {
-      fontStyle: 'underline',
-    },
-  };
-});
-
+import {DiscussionThreadItemActions} from './DiscussionThreadItemActions';
+import {DiscussionThreadItemFooter} from './DiscussionThreadItemFooter';
+import {MultiLineContent} from '../MultiLineContent/MultiLineContent';
 
 export const DiscussionThreadItem: FC<PropsWithChildren<DiscussionThreadItemProps>> = (props) => {
   const { root, preview, item, isSelected, pinned, as = 'li' } = props;
@@ -48,11 +30,9 @@ export const DiscussionThreadItem: FC<PropsWithChildren<DiscussionThreadItemProp
         >
           <Flex width="100%" justifyContent="space-between" marginTop="6px">
             <Flex grow={1} alignItems="center">
-              <Typography variant="omega" textColor="neutral800">
-                {/* TODO dangerouslySetInnerHTML in case of comments created by users is not safe
-                and can cause XSS attacks. We need to sanitize the content before displaying it. */}
-                <DiscussionThreadItemContentTypographyRenderer dangerouslySetInnerHTML={{ __html: item.content }} />
-              </Typography>
+              <MultiLineContent>
+                { item.content }
+              </MultiLineContent>
             </Flex>
             {!preview && (
               <DiscussionThreadItemActions
