+++ admin/src/pages/App/index.tsx
@@ -19,7 +19,7 @@ const InnerApp = () => {
   return (
     <Layouts.Root sideNav={<SideNav />}>
       <Routes>
-        <Route path="/discover" element={<CustomDiscover config={config} />} />
+        <Route path="/discover" element={<Discover config={config} />} />
         <Route path="/discover/:id" element={<Details config={config} />} />
         <Route path="/reports" element={<Reports config={config} />} />
         <Route path="*" element={<Navigate to="discover" replace />} />
