+++ admin/src/pages/Details/DiscussionThread.tsx
@@ -9,7 +9,6 @@ import { LoadingIndicatorOverlay } from '../../components/LoadingIndicatorOverla
 import { AllowedActions } from '../../types';
 import { getMessage } from '../../utils';
 import { ModeratorResponse } from './ModeratorResponse';
-import {CustomDiscussionThreadItem} from '../../components/CustomDiscussionThreadItem';
 
 type DiscussionThreadProps = {
   readonly allowedActions: AllowedActions;
@@ -48,7 +47,7 @@ export const DiscussionThread: FC<DiscussionThreadProps> = ({ isReloading, level
           const isSelected = selected?.id === item.id;
           const isThreadAuthor = !isNil(selected?.threadOf?.author?.id) && selected?.threadOf?.author?.id === item?.author?.id;
           return (
-            <CustomDiscussionThreadItem
+            <DiscussionThreadItem
               key={`comment-${item.id}`}
               item={item as unknown as any}
               root={isNil(rootThread)}
