+++ admin/src/pages/Discover/index.tsx
@@ -2,20 +2,16 @@ import {Table, Tbody, Th, Thead, Tr, Typography} from '@strapi/design-system';
 import {Layouts, Page, Pagination, SearchInput, useQueryParams} from '@strapi/strapi/admin';
 import {FC} from 'react';
 import {Config} from '../../api/schemas';
-import {CommentRow} from '../../components/CommentRow';
 import {CommentsStatusFilters} from '../../components/CommentStatusFilters';
 import {SortableTh} from '../../components/SortableTh';
 import {useCommentsAll} from '../../hooks/useCommentsAll';
 import {getMessage} from '../../utils';
+import {CommentRow} from '../../components/CommentRow';
 
 const tableHeaders = [
   { label: "page.discover.table.header.id" },
   { label: "page.discover.table.header.author" },
-  { label: "page.discover.table.header.email" },
   { label: "page.discover.table.header.message", orderBy: "content" },
-  { label: "page.discover.table.header.thread" },
-  { label: "page.discover.table.header.entry" },
-  { label: "page.discover.table.header.rating", orderBy: "rating" },
   { label: "page.discover.table.header.createdAt", orderBy: "createdAt" },
   { label: "page.discover.table.header.status" },
 ];
@@ -55,20 +51,22 @@ export const Discover: FC<{ config: Config }> = ({ config }) => {
                 {tableHeaders.map(({ label, orderBy }) => (
                   <>
                     {!orderBy ? (
-                      <Th>
+                      <Th key={label}>
                         <Typography variant="sigma">
-                          {getMessage(label)}
+                          {getMessage(label, "")}
                         </Typography>
                       </Th>
                     ) : (
                       <SortableTh
-                        label={getMessage(label)}
+                        key={label}
+                        label={getMessage(label, "")}
                         orderByKey={orderBy}
                       />
                     )}
                   </>
                 ))}
                 <Th />
+                <Th></Th>
               </Tr>
             </Thead>
             <Tbody>
@@ -89,4 +87,3 @@ export const Discover: FC<{ config: Config }> = ({ config }) => {
     </>
   );
 };
-
