+++ package.json
@@ -50,6 +50,9 @@
     "url": "https://github.com/VirtusLab-Open-Source/strapi-plugin-comments"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
     "@tanstack/react-query": "^5.40.0",
     "@types/sanitize-html": "^2.13.0",
     "codemirror5": "npm:codemirror@^5.65.11",
